refactor: consolidate title bar components
